Move point_cloud to rviz_rendering
